requirements.txt
fastapi
uvicorn
httpx
python-dotenv
numpy
pandas
matplotlib
requests
beautifulsoup4
duckdb
seaborn
scikit-learn